requirements.txt
# Core Agent Framework
google-adk==1.18.0

# API & Deployment
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
httpx==0.27.2

# Observability & Monitoring
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0

# Data Processing
pandas==2.2.3
numpy==2.1.3

# Medical/Scientific Libraries
biopython==1.84  # For PubMed/biomedical data
requests==2.32.3

# Testing
pytest==8.4.2
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# Utilities
python-dotenv==1.0.1
tqdm==4.66.5
tabulate==0.9.0
deprecated==1.2.14

# ML/Evaluation
scikit-learn==1.5.2

# Development
black==24.10.0
ruff==0.7.3